Refactor Pages for more organization
