+++ src/pages/Notices/NoticeRegister/index.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react';
 
-import Header from '../../components/Header';
-import { AppButton, ContentView, Form, Page } from '../../styles/default';
-import Footer from '../../components/Footer';
-import api from '../../services/api';
+import Header from '../../../components/Header';
+import { AppButton, ContentView, Form, Page } from '../../../styles/default';
+import Footer from '../../../components/Footer';
+import api from '../../../services/api';
 
 function Tag(props) {
   return (
