+++ src/pages/Notices/NoticesDetails/index.js
@@ -1,13 +1,12 @@
 import React, { useState, useEffect } from 'react';
-import Header from '../../components/Header';
-import NoticesCard from '../../components/NoticesCard';
-import { Page, ScreenView } from '../../styles/default';
-import imgTest from '../../assets/icon.svg';
+import Header from '../../../components/Header';
+import NoticesCard from '../../../components/NoticesCard';
+import { Page, ScreenView } from '../../../styles/default';
 
-import api from '../../services/api';
+import api from '../../../services/api';
 
-import { isAdmin } from '../../services/auth';
-import Footer from '../../components/Footer';
+import { isAdmin } from '../../../services/auth';
+import Footer from '../../../components/Footer';
 
 function NoticesDetails(props) {
