+++ src/routes/routes.js
@@ -5,13 +5,13 @@ import NotFound from '../pages/NotFound';
 import history from '../services/history/history';
 import Login from '../pages/Login';
 import AnunciarServico from '../pages/AnunciarServico';
-import OpportunitieList from '../pages/OpportunitieList';
-import OpportunitieRegister from '../pages/OpportunitieRegister';
-import OpportunitieEdit from '../pages/OpportunitieEdit';
-import OpportunitiePendents from '../pages/OpportunitiePendents';
-import NoticeRegister from '../pages/NoticeRegister';
-import NoticesList from '../pages/NoticesList';
-import NoticesDetails from '../pages/NoticesDetails';
+import OpportunitieList from '../pages/Oportunities/OpportunitieList';
+import OpportunitieRegister from '../pages/Oportunities/OpportunitieRegister';
+import OpportunitieEdit from '../pages/Oportunities/OpportunitieEdit';
+import OpportunitiePendents from '../pages/Oportunities/OpportunitiePendents';
+import NoticeRegister from '../pages/Notices/NoticeRegister';
+import NoticesList from '../pages/Notices/NoticesList';
+import NoticesDetails from '../pages/Notices/NoticesDetails';
 import Dashboard from '../pages/Dashboard';
 import TagEdit from '../pages/TagEdit';
 import UserEdit from '../pages/UserEdit';
